pyproject.toml
[project]
name = "deescm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "jinja2>=3.1.6",
    "python-decouple>=3.8",
    "requests>=2.32.3",
]
